DOP-2196: Remove \x00 characters from text nodes
docutils replaces escape characters with \x00. Our two options
are to either monkey-patch docutils to prevent that (risky, brittle),
or we can undo this once we ingest docutils output.

docutils provides an unescape() function, but it does a lot of things
based on input parameters, and let's just not play that game.