Remove the entire PendingTask infrastructure
It's been vestigual at best since the postprocessor was added in prehistoric times,
but I think it's only adding conceptual overhead and One More Noun to the project
structure.

It's time.

There is a notable change here in that now the JSONVisitor will immediately evaluate
figures and raised diagnostics, rather than waiting for that future finish() call
to happen back in the main process.

Overall performance-wise, this change is negligable. If a the project repeats a given
static asset very frequently, the removal the of the expensive_operation_cache could
theoretically have an impact, but that just doesn't happen in practice.