DOP-4491: Ensure that cache data is invalidated when images change
This removes the expensive_operation_cache, which is very old. The purpose
is to avoid re-hashing image assets, but I don't think the benefit is worth
having this weird interface around.

We should revisit in a post free-threading Python world when we can actually
give workers access to shared state.